[FIX]Fixed to not manage interfaces in bulk.
