+++ CuRendCore/CuRendCore/CuRendCore.vcxproj
@@ -32,6 +32,7 @@
       <PrecompiledHeaderFile Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">CRC_pch.h</PrecompiledHeaderFile>
       <PrecompiledHeaderFile Condition="'$(Configuration)|$(Platform)'=='Release|x64'">CRC_pch.h</PrecompiledHeaderFile>
     </ClCompile>
+    <ClCompile Include="src\CRC_phase_method.cpp" />
     <ClCompile Include="src\CRC_scene.cpp" />
     <ClCompile Include="src\CRC_window.cpp" />
   </ItemGroup>
@@ -39,9 +40,9 @@
     <ClInclude Include="include\CRC_config.h" />
     <ClInclude Include="include\CRC_core.h" />
     <ClInclude Include="include\CRC_funcs.h" />
-    <ClInclude Include="include\CRC_interface.h" />
     <ClInclude Include="include\CRC_pch.h" />
     <ClInclude Include="include\CRC_container.h" />
+    <ClInclude Include="include\CRC_phase_method.h" />
     <ClInclude Include="include\CRC_scene.h" />
     <ClInclude Include="include\CRC_window.h" />
     <ClInclude Include="include\CuRendCore.h" />
