+++ CuRendCore/CuRendCore/include/CuRendCore.h
@@ -4,9 +4,9 @@
 
 #include "CRC_config.h"
 #include "CRC_core.h"
-#include "CRC_interface.h"
-
-#include "CRC_container.h"
 
 #include "CRC_window.h"
 #include "CRC_scene.h"
+
+#include "CRC_container.h"
+#include "CRC_phase_method.h"
