+++ CuRendCore/CuRendCore/src/CRC_core.cpp
@@ -5,6 +5,8 @@
 
 #include "CRC_window.h"
 #include "CRC_scene.h"
+#include "CRC_container.h"
+#include "CRC_phase_method.h"
 
 CRCCore::CRCCore()
 {
