+++ CuRendCore/CuRendCore/src/CRC_phase_method.cpp
@@ -1,0 +1,4 @@
+﻿#include "CRC_pch.h"
+#include "CRC_funcs.h"
+
+#include "CRC_phase_method.h"
